Jul 20, 2023, 4:43 PM

+++ Course 1/Week 2/Python Solution.py
@@ -21,7 +21,7 @@ def CountSplitInv(B,C):
 def Sort_Count(A):
     n = len(A)
     if n > 1:
-        splitposition = int(n / 2)
+        splitposition = (n / 2)
         B,X = Sort_Count(A[:-splitposition])
         C,Y = Sort_Count(A[-splitposition:])
         D,Z = CountSplitInv(B,C)
